requirements.txt
torch==2.2.1
torchvision==0.17.1
open_clip_torch==2.24.0
transformers==4.38.2
scikit-learn==1.4.1.post1
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
albumentations==1.4.1
pyyaml==6.0.1
rich==13.7.1
loguru==0.7.2
tqdm==4.66.2
kaggle==1.6.6
seaborn==0.13.2
reportlab==4.1.0
